requirements.txt
Django==1.9.8
django-treebeard==4.0.1
django-tables2==1.2.3
mysqlclient==1.3.7
gunicorn==19.6.0
whitenoise==3.2